WIB-61 Rename | 경로명 변경
